backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "chromadb>=1.3.5",
    "fastapi>=0.121.3",
    "google-generativeai>=0.8.5",
    "groq>=0.36.0",
    "markdown>=3.10",
    "numpy>=2.3.5",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.38.0",
]
